Merged commands "help" and "command"
Renamed command "command" to "help" so now there are two "help" commands differently overloaded.
Using just "command" without parameters still sends a DM but putting in a command name as parameter shows the usage of that command.

I think this is more intuitive to use and reduces the total number of commands.