Improved handling of commands to be acutally asynchronous
Problem was that `_service.ExecuteAsync(context, argPos)` got called with the keyword "await" which meant if a command was still being executed and another got triggerd, the program first had to finish execution of the current command and then was able to start with the next one.
By not using `await` but calling `.ContinueWith(...)` on the resulting Task, multiple commands can be executed simutaniously even if they would contain a `await Task.Delay(100000);` which before was not possible.