Added Miunies-Rewards for sending messages
Currently rewards sending a message that is longer than 20 chars with 1-5 Miunies.
This reward has a cooldown of 30 seconds individually for each User.

These values are configurable inside the Constants class.

NOTE:
Not sure if it is the way you like it...
It would be less overhead (and less asynchronious things happening) to just add the function call of HandleMessageRewards() inside the CommandHander.HandleCommandAsync() and not register it to the client.OnMessageReceived Event.
The checks are essencially the same but it is not really about handling a command... so I felt putting it there would just make the code more spaghetti than it needs to be ^^

My biggest worry is that since it is updating (loading, modifying and saving) a userAccount asynchroniously that we could possibly run into race conditions if other commands do that aswell...
Like if it wants to reward a user and loads the account but then gets interrupted by the async scheduler a command continues to get executed and saves something to the account - and we would have an invalid / old userAccount in memory and probably will overwrite the change that the command did to it later on...
I have tested it quite a bit and it never seemed to have happend but being aware of this is probably not bad.
I don't know... maybe you know more than me how to address that or if I am just being overcautious...