Dependency Injection & Code Refactoring
* **Now using Dependency Injection**
DiscordSocketClient, CommandService, CommandHandler and DiscordEventHandler are the services so far. Not all are used everywhere yet... Global.Client is still used in most cases and this should be changed to inject the DiscordSocketClient object instead. But it is a step in the right direction.
* **Add Class / Service: DiscordEventHandler**
While moving to DI I noticed that subscriptions to events are scattered all around and burried in code everywhere so I decided to create a service called DiscordEventHandler and to move all subscriptions to events there and move some event functions to classes where they belong (e.G. UserJoined / UserLeft from Program.cs to Announcements.cs & ReactionAdded from Program.cs to BlockHandler)
* **Cleanup of Program.cs**
Additional to the above noted cleanings inside Program.cs handling commandline arguments moved to a function.